+++ src/lib/helpers/parser.ts
@@ -1,9 +1,3 @@
-export const pushHistory = (params: string) => {
-	const currentURL = window.location.href;
-	const baseURL = currentURL.split('?')[0];
-	window.history.pushState(null, '', `${baseURL}${params}`);
-};
-
 import { XMLParser } from 'fast-xml-parser';
 
 export const parser = new XMLParser({ ignoreAttributes: false });
